+++ src/commands/mod.rs
@@ -1,3 +1,2 @@
-// src/commands/mod.rs
-
+/// Contains the main command logic for `cleansh`.
 pub mod cleansh;